+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.  
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).  
 
+
+## [0.2.3] – 2026-01-06
+
+### Added
+- Added option to pass `r_bins` to `Grid3D`
+
+### Fixed
+- Grid3D consistent with *physics* convention
+- Grid3D tests and tutorials are updated
+
+
 ## [0.2.2] – 2026-01-05
 
 ### Added